refactor: simplify since there is only one task
When ecs.RunTask is called with Count unset, then there should always be exactly 1 task. This commit simplifies the code under that assumption so that we do not need an array of tasks. Also I don't think some of the functions worked anyway with multiple tasks (e.g. waitExitTasks returned when the first task finished instead of the last task).